requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
argcomplete==3.1.2
asgiref==3.7.2
attrs==23.2.0
bcrypt==3.2.0
blinker==1.4
boto3==1.34.14
botocore==1.34.14
Brlapi==0.8.3
cachetools==5.3.0
certifi==2023.11.17
cfn-flip==1.3.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cmdstanpy==1.2.4
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-psdb-engine==1.0.6
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
duplicity==0.8.21
durationpy==0.5
ecdsa==0.18.0
enum-compat==0.0.3
exceptiongroup==1.2.1
fastapi==0.110.2
fasteners==0.14.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
future==0.18.2
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.11.0
google-api-python-client==2.79.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-generativeai==0.3.2
googleapis-common-protos==1.58.0
greenlet==1.1.2
grpcio==1.60.0
grpcio-status==1.60.0
guake==3.8.5
gunicorn==21.2.0
h11==0.14.0
hjson==3.1.0
holidays==0.58
httplib2==0.20.2
huggingface-hub==0.26.0
idna==3.6
importlib-metadata==4.6.4
importlib_resources==6.4.5
intuit-oauth==1.2.5
jeepney==0.7.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kappa==0.6.0
keyring==23.5.0
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mysqlclient==2.2.0
netifaces==0.11.0
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
packaging==23.2
pandas==2.0.3
paramiko==2.9.3
pbr==5.8.0
pexpect==4.8.0
Pillow==9.0.1
placebo==0.9.0
plotly==5.24.1
prophet==1.1.6
proto-plus==1.23.0
protobuf==4.22.0
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyarrow==14.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.8.1b0
PyDrive==1.3.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-slugify==8.0.1
pytz==2023.3.post1
pyxdg==0.27
PyYAML==6.0.1
referencing==0.32.1
regex==2024.9.11
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.5
scikit-learn==1.0.2
scipy==1.10.1
sdkmanager==0.5.1
SecretStorage==3.3.1
six==1.16.0
sklearn==0.0.post1
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==1.4.31
sqlparse==0.4.4
stanio==0.5.1
starlette==0.37.2
streamlit==1.30.0
stripe==7.11.0
systemd-python==234
tenacity==8.2.3
text-unidecode==1.3
threadpoolctl==3.1.0
tokenizers==0.20.1
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
transformers==4.45.2
troposphere==4.4.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.29.0
validators==0.22.0
wadllib==1.3.6
watchdog==3.0.0
Werkzeug==3.0.0
xdg==5
xgboost==1.7.4
xkit==0.0.0
zappa==0.58.0
zipp==1.0.0
